separate planning strategies from DevicesPlanner into DevicesPlannerStrategies module
There are currently :legacy and a stub for :ng. The strategy is read from proposal settings.
